server/pyproject.toml
[project]
name = "140-chatbot-fastapi-nextjs"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "NatsukiHayashi", email = "natsuki.hayashi@arsaga.jp" }
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.31.0",
    "sqlalchemy[asyncio]>=2.0.35",
    "alembic>=1.13.3",
    "asyncpg>=0.29.0",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.6.8",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/_140_chatbot_fastapi_nextjs"]
